// apps/lxp/modules/training-contract/main/training-tab/table/types.d.ts
export type TsubjectItems = {
  id: string;
  teacher_id: string;
  department_id: string;
  name: string;
  subject_code: string;
  degree: string;
  level: number;
  indicator: string;
  study_experience: string;
  teaching_materials: string;
  basic_competencies: string;
  tools_needed: string;
  scoring: string;
  description: string;
  thumbnail: string;
  credit: number;
  session_total_number: number;
  slug: string;
};